wip: init module
